Moves flash message logic to concern.
